rename instance_of.ts to instance-of.ts
