+++ src/types/algorand-client-interface.ts
@@ -1,5 +1,5 @@
 import { AlgorandClient } from './algorand-client'
-import { InterfaceOf } from './instance_of'
+import { InterfaceOf } from './instance-of'
 
- /** AlgorandClient interface used to eleminate circular dependencies */
+/** AlgorandClient interface used to eleminate circular dependencies */
 export type AlgorandClientInterface = InterfaceOf<AlgorandClient>
