+++ src/types/instance-of.ts
